Split logical + server into module
